Move StringViewSaver to common.h
